requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# GitHub integration
PyGithub>=2.1.1
requests>=2.31.0

# YAML processing
PyYAML>=6.0.1
ruamel.yaml>=0.18.5

# Docker for sandbox validation
docker>=7.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
httpx>=0.25.0

# Logging and monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1
types-PyYAML>=6.0.12
types-requests>=2.31.0
